+++ packages/vc/src/__tests__/core/vc.test.ts
@@ -1,0 +1,375 @@
+import Ajv, { ValidateFunction } from 'ajv'
+import addFormats from 'ajv-formats'
+
+import { VCProof, vcProofSchema, VCType, vcTypeSchema } from '../../core'
+
+describe('vcProofSchema', () => {
+  let ajv: Ajv
+  let validate: ValidateFunction<VCProof>
+
+  beforeAll(() => {
+    ajv = addFormats(new Ajv())
+    ajv.addSchema(vcProofSchema, 'vcProofSchema')
+  })
+
+  beforeEach(() => {
+    validate = ajv.getSchema('vcProofSchema')!
+  })
+
+  test('valiates a valid VC Proof', () => {
+    const value: VCProof = {
+      type: 'EcdsaSecp256k1Signature2019',
+      created: '2021-06-03T00:00:00Z',
+      proofPurpose: 'assertionMethod',
+      verificationMethod: 'did:example:issuer#123',
+      jws: 'jws',
+    }
+
+    expect(validate(value)).toBeTruthy()
+  })
+
+  describe('fails an invalid VC Proof:', () => {
+    test('missing type', () => {
+      const value = {
+        created: '2021-06-03T00:00:00Z',
+        proofPurpose: 'assertionMethod',
+        verificationMethod: 'did:example:issuer#123',
+        jws: 'jws',
+      }
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'type'",
+            "params": Object {
+              "missingProperty": "type",
+            },
+            "schemaPath": "#/oneOf/0/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'type'",
+            "params": Object {
+              "missingProperty": "type",
+            },
+            "schemaPath": "#/oneOf/1/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "oneOf",
+            "message": "must match exactly one schema in oneOf",
+            "params": Object {
+              "passingSchemas": null,
+            },
+            "schemaPath": "#/oneOf",
+          },
+        ]
+      `)
+    })
+
+    test('missing created', () => {
+      const value = {
+        type: 'EcdsaSecp256k1Signature2019',
+        proofPurpose: 'assertionMethod',
+        verificationMethod: 'did:example:issuer#123',
+        jws: 'jws',
+      }
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'created'",
+            "params": Object {
+              "missingProperty": "created",
+            },
+            "schemaPath": "#/oneOf/0/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'created'",
+            "params": Object {
+              "missingProperty": "created",
+            },
+            "schemaPath": "#/oneOf/1/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "oneOf",
+            "message": "must match exactly one schema in oneOf",
+            "params": Object {
+              "passingSchemas": null,
+            },
+            "schemaPath": "#/oneOf",
+          },
+        ]
+      `)
+    })
+
+    test('missing proofPurpose', () => {
+      const value = {
+        type: 'EcdsaSecp256k1Signature2019',
+        created: '2021-06-03T00:00:00Z',
+        verificationMethod: 'did:example:issuer#123',
+        jws: 'jws',
+      }
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'proofPurpose'",
+            "params": Object {
+              "missingProperty": "proofPurpose",
+            },
+            "schemaPath": "#/oneOf/0/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'proofPurpose'",
+            "params": Object {
+              "missingProperty": "proofPurpose",
+            },
+            "schemaPath": "#/oneOf/1/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "oneOf",
+            "message": "must match exactly one schema in oneOf",
+            "params": Object {
+              "passingSchemas": null,
+            },
+            "schemaPath": "#/oneOf",
+          },
+        ]
+      `)
+    })
+
+    test('invalid proofPurpose', () => {
+      const value = {
+        type: 'EcdsaSecp256k1Signature2019',
+        created: '2021-06-03T00:00:00Z',
+        proofPurpose: 'authentication',
+        verificationMethod: 'did:example:issuer#123',
+        jws: 'jws',
+      }
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "/proofPurpose",
+            "keyword": "const",
+            "message": "must be equal to constant",
+            "params": Object {
+              "allowedValue": "assertionMethod",
+            },
+            "schemaPath": "#/oneOf/0/properties/proofPurpose/const",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'proofValue'",
+            "params": Object {
+              "missingProperty": "proofValue",
+            },
+            "schemaPath": "#/oneOf/1/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "oneOf",
+            "message": "must match exactly one schema in oneOf",
+            "params": Object {
+              "passingSchemas": null,
+            },
+            "schemaPath": "#/oneOf",
+          },
+        ]
+      `)
+    })
+
+    test('missing verificationMethod', () => {
+      const value = {
+        type: 'EcdsaSecp256k1Signature2019',
+        created: '2021-06-03T00:00:00Z',
+        proofPurpose: 'assertionMethod',
+        jws: 'jws',
+      }
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'verificationMethod'",
+            "params": Object {
+              "missingProperty": "verificationMethod",
+            },
+            "schemaPath": "#/oneOf/0/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'verificationMethod'",
+            "params": Object {
+              "missingProperty": "verificationMethod",
+            },
+            "schemaPath": "#/oneOf/1/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "oneOf",
+            "message": "must match exactly one schema in oneOf",
+            "params": Object {
+              "passingSchemas": null,
+            },
+            "schemaPath": "#/oneOf",
+          },
+        ]
+      `)
+    })
+
+    test('missing jws', () => {
+      const value = {
+        type: 'EcdsaSecp256k1Signature2019',
+        created: '2021-06-03T00:00:00Z',
+        proofPurpose: 'assertionMethod',
+        verificationMethod: 'did:example:issuer#123',
+      }
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'jws'",
+            "params": Object {
+              "missingProperty": "jws",
+            },
+            "schemaPath": "#/oneOf/0/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "required",
+            "message": "must have required property 'proofValue'",
+            "params": Object {
+              "missingProperty": "proofValue",
+            },
+            "schemaPath": "#/oneOf/1/required",
+          },
+          Object {
+            "instancePath": "",
+            "keyword": "oneOf",
+            "message": "must match exactly one schema in oneOf",
+            "params": Object {
+              "passingSchemas": null,
+            },
+            "schemaPath": "#/oneOf",
+          },
+        ]
+      `)
+    })
+  })
+})
+
+describe('vcTypeSchema', () => {
+  let ajv: Ajv
+  let validate: ValidateFunction<VCType>
+
+  beforeAll(() => {
+    ajv = addFormats(new Ajv())
+    ajv.addSchema(vcTypeSchema, 'vcTypeSchema')
+  })
+
+  beforeEach(() => {
+    validate = ajv.getSchema('vcTypeSchema')!
+  })
+
+  describe('valiates a valid VC Type:', () => {
+    test('single item', () => {
+      const value: VCType = ['VerifiableCredential']
+
+      expect(validate(value)).toBeTruthy()
+    })
+
+    test('multiple item', () => {
+      const value: VCType = ['VerifiableCredential', 'CustomCredential']
+
+      expect(validate(value)).toBeTruthy()
+    })
+
+    test('multiple item with different order', () => {
+      const value: VCType = ['CustomCredential', 'VerifiableCredential']
+
+      expect(validate(value)).toBeTruthy()
+    })
+  })
+
+  describe('fails an invalid VC Type:', () => {
+    test('empty array', () => {
+      const value: string[] = []
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "minItems",
+            "message": "must NOT have fewer than 1 items",
+            "params": Object {
+              "limit": 1,
+            },
+            "schemaPath": "#/minItems",
+          },
+        ]
+      `)
+    })
+
+    test('missing VerifiableCredential', () => {
+      const value: string[] = ['CustomCredential']
+
+      const result = validate(value)
+
+      expect(result).toBeFalsy()
+      expect(validate.errors).toMatchInlineSnapshot(`
+        Array [
+          Object {
+            "instancePath": "",
+            "keyword": "contains",
+            "message": "must contain at least 1 valid item(s)",
+            "params": Object {
+              "minContains": 1,
+            },
+            "schemaPath": "#/contains",
+          },
+        ]
+      `)
+    })
+  })
+})