+++ packages/vc/src/__tests__/sign.test.ts
@@ -43,8 +43,6 @@ describe('signVC', () => {
         suite,
       })
 
-      console.log({ proof: signed.proof })
-
       expect(validate(signed)).toBeTruthy()
       expectType<VC>(signed)
     })