+++ packages/vc/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+- Loosen schema for "type" arrays
+  - Now allows "VerifiableCredential"/"VerifiablePresentation" anywhere in the array instead of enforcing it as the first item
+
 ## 0.1.2
 
 - Improve the `RemoveIndex` utility type